+++ module/item/item-attack.js
@@ -1239,7 +1239,7 @@ export async function _onApplyDamageToSpecificToken(event, tokenId) {
 
 
 
-    const avad = item.findModsByXmlid( "AVAD");
+    const avad = item.findModsByXmlid("AVAD");
 
     // Check for conditional defenses
     let ignoreDefenseIds = []
@@ -1439,15 +1439,39 @@ export async function _onApplyDamageToSpecificToken(event, tokenId) {
         return _onApplySenseAffectingToSpecificToken(event, tokenId, damageData, defense)
     }
 
+    // AUTOMATION immune to mental powers
+    if (item.system.class === "mental" && token?.actor?.type === "automaton") {
+        defenseTags.push({
+            "name": "AUTOMATON",
+            "value": "immune",
+            "resistant": false,
+            "title": "Automations are immune to mental powers",
+        })
+        damageDetail.stun = 0
+        damageDetail.body = 0
+    }
+
     // AUTOMATION powers related to STUN
     const CANNOTBESTUNNED = token.actor.items.find(o => o.system.XMLID === "AUTOMATON" && o.system.OPTION === "CANNOTBESTUNNED")
     const NOSTUN1 = token.actor.items.find(o => o.system.XMLID === "AUTOMATON" && o.system.OPTION === "NOSTUN1") // AUTOMATION Takes No STUN (loses abilities when takes BODY)
     const NOSTUN2 = token.actor.items.find(o => o.system.XMLID === "AUTOMATON" && o.system.OPTION === "NOSTUN2") //Takes No STUN
-    if (NOSTUN1) {
+    if (NOSTUN1 && damageDetail.stun > 0) {
+        defenseTags.push({
+            "name": "TAKES NO STUN",
+            "value": "immune",
+            "resistant": false,
+            "title": "Ignore the STUN damage from any attack; loses abilities when takes BODY",
+        })
         damageDetail.effects = damageDetail.effects + "Takes No STUN (loses abilities when takes BODY); "
         damageDetail.stun = 0;
     }
-    if (NOSTUN2) {
+    if (NOSTUN2 && damageDetail.stun > 0) {
+        defenseTags.push({
+            "name": "TAKES NO STUN",
+            "value": "immune",
+            "resistant": false,
+            "title": "Ignore the STUN damage from any attack",
+        })
         damageDetail.effects = damageDetail.effects + "Takes No STUN; "
         damageDetail.stun = 0;
     }
@@ -1482,7 +1506,7 @@ export async function _onApplyDamageToSpecificToken(event, tokenId) {
             }
         }
     }
-    effectsFinal = effectsFinal.replace(/; $/,"")
+    effectsFinal = effectsFinal.replace(/; $/, "")
 
     let cardData = {
         item: item,
@@ -1565,7 +1589,7 @@ async function _onApplyAdjustmentToSpecificToken(event, tokenId, damageData, def
     }
 
     if (item.actor.id === token.actor.id && ["DRAIN", "TRANSFER"].includes(item.system.XMLID)) {
-        return ui.notifications.error(`${item.system.XMLID} attacker (${item.actor.name}) and defender (${token.actor.name}) cannot be the same.`);
+        await ui.notifications.warn(`${item.system.XMLID} attacker (${item.actor.name}) and defender (${token.actor.name}) are the same.`);
     }
 
 
@@ -1914,7 +1938,11 @@ async function _calcDamage(damageResult, item, options) {
 
     let hasStunMultiplierRoll = false;
     //let renderedStunMultiplierRoll = null;
-    let stunMultiplier = 1;
+
+    const INCREASEDSTUNMULTIPLIER = item.findModsByXmlid("INCREASEDSTUNMULTIPLIER")
+
+    let stunMultiplier = 1 + parseInt(INCREASEDSTUNMULTIPLIER?.LEVELS || 0)
+
     let noHitLocationsPower = item.system.noHitLocations || false;
 
 
@@ -1994,6 +2022,8 @@ async function _calcDamage(damageResult, item, options) {
         } else {
             stunMultiplier = stunResult.total;
         }
+
+        stunMultiplier += parseInt(INCREASEDSTUNMULTIPLIER?.LEVELS || 0)
 
         if (options.stunmultiplier) {
             stunMultiplier = options.stunmultiplier